# ebay-service-spring/ebay-service/src/main/java/soa/duo/ebayservice/model/Coordinates.java
package soa.duo.ebayservice.model;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class Coordinates {

    private Integer x;
    private int y;

}